UI: Widen compatibility list, and make search box take up all space horizontally
