+++ src/Ryujinx/Utilities/Compat/CompatibilityContentDialog.axaml
@@ -1,0 +1,20 @@
+﻿<ui:ContentDialog xmlns="https://github.com/avaloniaui"
+                  xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+                  xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+                  xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+                  xmlns:local="using:Ryujinx.Ava.Utilities.Compat"
+                  xmlns:ui="using:FluentAvalonia.UI.Controls"
+                  xmlns:ext="using:Ryujinx.Ava.Common.Markup"
+                  x:Class="Ryujinx.Ava.Utilities.Compat.CompatibilityContentDialog"
+                  mc:Ignorable="d" d:DesignWidth="600" d:DesignHeight="400"
+                  CloseButtonText="{ext:Locale SettingsButtonClose}" 
+                  DefaultButton="Close"
+                  x:DataType="local:CompatibilityViewModel">
+    <ui:ContentDialog.DataContext>
+        <local:CompatibilityViewModel/>
+    </ui:ContentDialog.DataContext>
+    <ui:ContentDialog.Resources>
+        <x:Double x:Key="ContentDialogMaxWidth">900</x:Double>
+    </ui:ContentDialog.Resources>
+</ui:ContentDialog>
+
