+++ src/Ryujinx/Utilities/Compat/CompatibilityList.axaml
@@ -11,25 +11,24 @@
     <UserControl.DataContext>
         <local:CompatibilityViewModel />
     </UserControl.DataContext>
-    <StackPanel Orientation="Vertical">
-        <StackPanel HorizontalAlignment="Right" Orientation="Horizontal">
-            <TextBox Watermark="{ext:Locale CompatibilityListSearchBoxWatermark}" TextChanged="TextBox_OnTextChanged" />
-            <CheckBox Margin="7, 0, 0, 0" IsChecked="{Binding OnlyShowOwnedGames}" />
-            <TextBlock Margin="-3, 0, 0, 0" Text="{ext:Locale CompatibilityListOnlyShowOwnedGames}" />
-        </StackPanel>
-        <ScrollViewer>
-                <ListBox Margin="5"
-                         Padding="10"
+    <Grid RowDefinitions="Auto,*">
+        <Grid Grid.Row="0" ColumnDefinitions="*,Auto,Auto">
+            <TextBox Grid.Column="0" HorizontalAlignment="Stretch" Watermark="{ext:Locale CompatibilityListSearchBoxWatermark}" TextChanged="TextBox_OnTextChanged" />
+            <CheckBox Grid.Column="1" Margin="7, 0, 0, 0" IsChecked="{Binding OnlyShowOwnedGames}" />
+            <TextBlock Grid.Column="2" Margin="-10, 0, 0, 0" Text="{ext:Locale CompatibilityListOnlyShowOwnedGames}" />
+        </Grid>
+        <ScrollViewer Grid.Row="1">
+                <ListBox Margin="0,5, 0, 0"
                          Background="Transparent"
                          ItemsSource="{Binding CurrentEntries}">
                     <ListBox.ItemTemplate>
                         <DataTemplate DataType="{x:Type local:CompatibilityEntry}">
-                            <Grid HorizontalAlignment="Center" Width="500" ColumnDefinitions="Auto,Auto,Auto,*"
+                            <Grid Width="750" ColumnDefinitions="Auto,Auto,Auto,*"
                                   Margin="5">
                                 <TextBlock Grid.Column="0"
                                            FontFamily="{StaticResource JetBrainsMono}"
                                            Text="{Binding GameName}"
-                                           Width="140"
+                                           Width="333"
                                            TextWrapping="Wrap" />
                                 <TextBlock Grid.Column="1"
                                            Width="135"
@@ -55,5 +54,5 @@
                     </ListBox.ItemTemplate>
                 </ListBox>
         </ScrollViewer>
-    </StackPanel>
+    </Grid>
 </UserControl>
